implement efax
1. eFax page
i) Remove all the tab from the nav and only add "All" and "File"
ii) On the all section all the eFax history are shown
iii) On the File section all the available file name, view and download, and delete options.
iv) Implement file view section on the left side.
v) Implement a file upload page with all the functionality.