Responsive fix
If the user is zooming out, then the Device Pixel Ratio changes and browser misinterprets the Device Resolution regardless of screen size. So I am changing it to "device-width" so that it takes the screen resolution as ultimate. Do remember that Zooming In or Zooming Out will not feel the same as it used to.